Add color palette to storybook docs Remove the old src/docs removed a related dependency Remove unused icons from storybook docs static files
