+++ .storybook/lucid-docs-addon/PropTypes.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOMServer from 'react-dom/server';
 import _ from 'lodash';
 import marksy from 'marksy/components';
-import { stripIndent } from '../../src/docs/util';
+import { stripIndent } from './util';
 
 import SyntaxHighlighter, {
 	registerLanguage,
