+++ .storybook/manager-head.html
@@ -8,4 +8,5 @@
     font-family: "Helvetica Neue", Helvetica, Arial, Sans-Serif;
   }
 </style>
-<link rel="icon" type="image/png" href="/lucid/lucid-icon.png" />
+<link rel="icon" type="image/png" href="//appnexus.github.io/lucid/lucid-icon.png" />
+<link rel="icon" type="image/png" href="/lucid-icon.png" />
