+++ docs/color-palette.less
@@ -1,3 +1,5 @@
+@import '../src/styles/variables.less';
+
 .ColorPalette {
 	padding: 15px;
