--- docs/public/lucid/lucid-icon.png
@@ -1,1 +1,0 @@
-../lucid-icon.png
