+++ package.json
@@ -96,7 +96,6 @@
     "less": "^2.3.1",
     "less-loader": "^4.1.0",
     "lint-staged": "^3.4.0",
-    "markdown": "^0.5.0",
     "marksy": "^6.0.3",
     "mini-css-extract-plugin": "^0.4.0",
     "np": "^2.20.1",
