--- src/docs/containers/icons.less
@@ -1,32 +1,0 @@
-.Icons {
-	padding: 15px;
-
-	&-list {
-		padding: 0;
-		height: 75vh;
-		display: flex;
-		align-items: flex-start;
-		flex-flow: column wrap;
-		align-content: flex-start;
-	}
-
-	&-list-item {
-		list-style-type: none;
-		margin: @size-S;
-	}
-
-	&-list-icon {
-		margin-right: @size-S;
-	}
-
-	&-buttons {
-		.lucid-Button {
-			margin-right: @size-XXS;
-		}
-	}
-
-	&-buttons-section {
-		margin-bottom: @size-XL;
-	}
-}
-
