--- src/docs/containers/landing-page.less
@@ -1,44 +1,0 @@
-.lucid-LandingPage {
-	text-align: center;
-
-	&-logo {
-		display: none;
-
-		@media @large {
-			max-width: 100%;
-			display: initial;
-		}
-	}
-
-	&-primary {
-		fill: @color-primary;
-	}
-
-	&-readme {
-		text-align: left;
-		padding: @size-L;
-	}
-
-	&-carpetize {
-		cursor: pointer;
-	}
-
-	&-is-carpeted {
-		position: relative;
-		min-height: 100vh;
-
-		&::before {
-			content: '';
-			position: absolute;
-			top: 0;
-			left: 0;
-			bottom: 0;
-			right: 0;
-			background-image: data-uri('../img/pdx.svg');
-			background-size: 100px;
-			opacity: 0.1;
-			z-index: -1;
-		}
-	}
-}
-
