--- src/docs/index.html
@@ -1,15 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta http-equiv="content-type" content="text/html; charset=UTF8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Lucid</title>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.0.0/highlight.min.js"></script>
-    <link href="https://cdnjs.cloudflare.com/ajax/libs/highlight.js/9.0.0/styles/solarized-light.min.css" rel="stylesheet" type="text/css" />
-  </head>
-  <body>
-    <div id="docs">
-    </div>
-    <script src="./bundle.js"></script>
-  </body>
-</html>
